+++ app/src/routes/AppRouter.tsx
@@ -12,33 +12,39 @@ export const AppRouter = () => {
   const { authContext } = useContext(UserContext)
 
   return (
-      <Box d={'flex'} alignItems={'center'} flexDirection={'column'} minH={'100vh'}>
-        <Router>
-          <Box w='100%'>
-            <Navbar />
-          </Box>
-          <Box h='100%' width='100%' d='flex' flexGrow={1} flexDirection={'column'} bg={useColorModeValue('gray.100', 'gray.800')}>
-              { authContext.isLoading
-                ? <LoadingView />
-                : (
-                  <Switch>
-                  {authContext.isLoggedIn &&
-                  Object.keys(protectedRoutes).map(k => (
-                    authContext.isLoggedIn
-                    ? <Route key={k} exact path={protectedRoutes[k].path} component={protectedRoutes[k].component} />
-                    : <Redirect key={k} to={publicRoutes.login.path} />
-                  ))}
+    <Box d={'flex'} alignItems={'center'} flexDirection={'column'} minH={'100vh'}>
+      <Router>
+        <Box w='100%'>
+          <Navbar />
+        </Box>
+        <Box h='100%' width='100%' d='flex' flexGrow={1} flexDirection={'column'} bg={useColorModeValue('gray.100', 'gray.800')}>
+          { authContext.isLoading
+            ? <LoadingView />
+            : (
+              <Switch>
+                {Object.keys(publicRoutes).map(k => (
+                  <Route key={k} exact={k !== '404'} path={publicRoutes[k].path} component={publicRoutes[k].component}/>  
+                ))}
 
-                  {Object.keys(publicRoutes).map(k => (
-                    <Route key={k} exact={k !== '404'} path={publicRoutes[k].path} component={publicRoutes[k].component}/>
-                  ))}
-                  <Redirect to={publicRoutes.notFound.path} />
-                  </Switch>
-                )
-              }
-            <Footer />
-          </Box>
-        </Router>
-      </Box>
+                {Object.keys(protectedRoutes).map(k => (
+                  <Route key={k} exact
+                    path={protectedRoutes[k].path}
+                    render= { () => {
+                      const Component = protectedRoutes[k].component;
+                      return authContext.isLoggedIn
+                      ? <Component /> 
+                      : <Redirect to={publicRoutes.login.path}/>
+                    }}
+                  />
+                ))
+                }
+                <Redirect to={publicRoutes.notFound.path} />
+              </Switch>
+            )
+          }
+          <Footer />
+        </Box>
+      </Router>
+    </Box>
   )
 }
