+++ app/src/views/Conditions/ConditionsTable.tsx
