+++ app/src/views/Conditions/NewConditionForm.tsx
