+++ app/src/views/Navbar/Navbar.tsx
@@ -18,7 +18,6 @@ import {
   import { ColorModeSwitcher } from "../../theme/ColorModeSwitcher";
   import { UserContext } from '../../provider/AuthProvider';
 import { routes } from "../../routes/routes";
-import { userTypes } from '../../utils/typesDefinitions';
 
 type NavItem = {
   label: string;
@@ -40,10 +39,6 @@ export const Navbar = () => {
 
     if (authContext.isLoggedIn) {
       setHomePath(routes.home.path);
-      if(authContext.type?.includes(userTypes.MEDIC)){
-        setHomePath(routes.medic.path)
-        loggedInItems.push({label:'Mis condiciones', path:routes.home.path})
-      }
       setNavbarItems(loggedInItems);
     }else{
       setHomePath(routes.default.path)
